[MOV] Rename the folder fields to tables, which is more accurate.
