+++ libs/database/tables/discussion.h
