+++ libs/database/tables/thread.c
@@ -26,4 +26,7 @@ thread_t *thread_init(id_t id, id_t start_msg_id)
 
 void thread_destruction(thread_t *thread)
 {
+    if (!thread)
+        return;
+    free(thread);
 }
